refactor into services and types
